Use documented and capitalized methods in device.go ...
and have those methods call equally named but uncapitalized
methods in the platform specific device_darwin.go and
device_linux.go files
